Refactored code to be completely stateless
- On road to use Promises on the project
- Services are stateless
