--- DependencyInjection/CompilerPass/AppRepositoriesCompilerPass.php
@@ -1,54 +1,0 @@
-<?php
-
-/*
- * This file is part of the Apisearch Server
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- *
- * Feel free to edit as you please, and have fun.
- *
- * @author Marc Morera <yuhu@mmoreram.com>
- */
-
-declare(strict_types=1);
-
-namespace Apisearch\Server\DependencyInjection\CompilerPass;
-
-use Mmoreram\BaseBundle\CompilerPass\TagCompilerPass;
-
-/**
- * Class AppRepositoriesCompilerPass.
- */
-class AppRepositoriesCompilerPass extends TagCompilerPass
-{
-    /**
-     * Get collector service name.
-     *
-     * @return string Collector service name
-     */
-    public function getCollectorServiceName(): string
-    {
-        return 'apisearch_server.app_repository';
-    }
-
-    /**
-     * Get collector method name.
-     *
-     * @return string Collector method name
-     */
-    public function getCollectorMethodName(): string
-    {
-        return 'addRepository';
-    }
-
-    /**
-     * Get tag name.
-     *
-     * @return string Tag name
-     */
-    public function getTagName(): string
-    {
-        return 'apisearch_server.app_repository';
-    }
-}
