--- Domain/Repository/Repository/IndexRepository.php
@@ -1,31 +1,0 @@
-<?php
-
-/*
- * This file is part of the Apisearch Server
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- *
- * Feel free to edit as you please, and have fun.
- *
- * @author Marc Morera <yuhu@mmoreram.com>
- */
-
-declare(strict_types=1);
-
-namespace Apisearch\Server\Domain\Repository\Repository;
-
-use Apisearch\Model\Item;
-
-/**
- * Interface IndexRepository.
- */
-interface IndexRepository
-{
-    /**
-     * Generate items documents.
-     *
-     * @param Item[] $items
-     */
-    public function addItems(array $items);
-}
