--- Domain/Repository/Repository/UpdateRepository.php
@@ -1,36 +1,0 @@
-<?php
-
-/*
- * This file is part of the Apisearch Server
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- *
- * Feel free to edit as you please, and have fun.
- *
- * @author Marc Morera <yuhu@mmoreram.com>
- */
-
-declare(strict_types=1);
-
-namespace Apisearch\Server\Domain\Repository\Repository;
-
-use Apisearch\Model\Changes;
-use Apisearch\Query\Query;
-
-/**
- * Interface UpdateRepository.
- */
-interface UpdateRepository
-{
-    /**
-     * Update items.
-     *
-     * @param Query   $query
-     * @param Changes $changes
-     */
-    public function updateItems(
-        Query $query,
-        Changes $changes
-    );
-}
