+++ Domain/WithAppRepositoryAndEventPublisher.php
@@ -15,8 +15,8 @@ declare(strict_types=1);
 
 namespace Apisearch\Server\Domain;
 
-use Apisearch\App\AppRepository;
 use Apisearch\Server\Domain\Event\EventPublisher;
+use Apisearch\Server\Domain\Repository\AppRepository\Repository as AppRepository;
 
 /**
  * Class WithAppRepositoryAndEventPublisher.
