+++ Domain/WithRepositoryAndEventPublisher.php
@@ -15,8 +15,8 @@ declare(strict_types=1);
 
 namespace Apisearch\Server\Domain;
 
-use Apisearch\Repository\Repository;
 use Apisearch\Server\Domain\Event\EventPublisher;
+use Apisearch\Server\Domain\Repository\Repository\Repository;
 
 /**
  * Class WithRepositoryAndEventPublisher.
