+++ Plugin/Elastica/Resources/config/domain.yml
@@ -14,70 +14,47 @@ services:
     apisearch_plugin.elastica.result_builder:
         class: Apisearch\Plugin\Elastica\Domain\Builder\ResultBuilder
 
-    apisearch_plugin.elastica.item_elastica_wrapper:
-        class: Apisearch\Plugin\Elastica\Domain\ItemElasticaWrapper
+    apisearch_plugin.elastica.elastica_wrapper:
+        class: Apisearch\Plugin\Elastica\Domain\ElasticaWrapper
         arguments:
             - "@apisearch_plugin.elastica.elastica_client"
 
     #
     # Elastica repositories
     #
-    apisearch_plugin.elastica.elastica_wrapper_with_repository_reference:
-        class: Apisearch\Plugin\Elastica\Domain\ElasticaWrapperWithRepositoryReference
+
+
+    apisearch_plugin.elastica.with_elastica_wrapper:
+        class: Apisearch\Plugin\Elastica\Domain\WithElasticaWrapper
         abstract: true
         public: false
         arguments:
-            - "@apisearch_plugin.elastica.item_elastica_wrapper"
+            - "@apisearch_plugin.elastica.elastica_wrapper"
             - "%apisearch_plugin.elastica.refresh_on_write%"
 
-    apisearch_plugin.elastica.item_elastica_with_app_id_wrapper:
-        class: Apisearch\Plugin\Elastica\Domain\ElasticaWrapperWithRepositoryReference
-        parent: apisearch_plugin.elastica.elastica_wrapper_with_repository_reference
-        abstract: true
-        public: false
 
-    apisearch_plugin.elastica.elastica_query_repository:
+    #
+    # Elastica Items repositories
+    #
+    apisearch_server.item_repository:
+        class: Apisearch\Plugin\Elastica\Domain\Repository\ItemRepository
+        parent: apisearch_plugin.elastica.with_elastica_wrapper
+        arguments:
+            - "@apisearch_plugin.elastica.query_builder"
+
+    apisearch_server.query_repository:
         class: Apisearch\Plugin\Elastica\Domain\Repository\QueryRepository
-        parent: apisearch_plugin.elastica.item_elastica_with_app_id_wrapper
+        parent: apisearch_plugin.elastica.with_elastica_wrapper
         arguments:
             - "@apisearch_plugin.elastica.query_builder"
             - "@apisearch_plugin.elastica.result_builder"
-        tags:
-            - {name: "apisearch_server.items_repository"}
-
-    apisearch_plugin.elastica.update_repository:
-        class: Apisearch\Plugin\Elastica\Domain\Repository\UpdateRepository
-        parent: apisearch_plugin.elastica.item_elastica_with_app_id_wrapper
-        arguments:
-            - "@apisearch_plugin.elastica.query_builder"
-        tags:
-            - {name: "apisearch_server.items_repository"}
-
-    apisearch_plugin.elastica.elastica_delete_repository:
-        class: Apisearch\Plugin\Elastica\Domain\Repository\DeleteRepository
-        parent: apisearch_plugin.elastica.item_elastica_with_app_id_wrapper
-        tags:
-            - {name: "apisearch_server.items_repository"}
-
-    apisearch_plugin.elastica.elastica_index_repository:
-        class: Apisearch\Plugin\Elastica\Domain\Repository\IndexRepository
-        parent: apisearch_plugin.elastica.item_elastica_with_app_id_wrapper
-        tags:
-            - {name: "apisearch_server.items_repository"}
 
 
     #
     # Elastica App repositories
     #
-
-    apisearch_plugin.elastica.app_elastica_with_app_id_wrapper:
-        class: Apisearch\Plugin\Elastica\Domain\ElasticaWithAppIdWrapper
-        parent: apisearch_plugin.elastica.elastica_wrapper_with_repository_reference
-        abstract: true
-        public: false
-
-    apisearch_plugin.elastica.elastica_index_app_repository:
+    apisearch_server.index_repository:
         class: Apisearch\Plugin\Elastica\Domain\AppRepository\IndexRepository
-        parent: apisearch_plugin.elastica.app_elastica_with_app_id_wrapper
+        parent: apisearch_plugin.elastica.with_elastica_wrapper
         tags:
             - {name: "apisearch_server.app_repository"}
