+++ Tests/Functional/ApisearchServerBundleFunctionalTest.php
@@ -15,6 +15,7 @@ declare(strict_types=1);
 
 namespace Apisearch\Server\Tests\Functional;
 
+use Apisearch\ApisearchBundle;
 use Apisearch\Config\Config;
 use Apisearch\Exception\ResourceNotAvailableException;
 use Apisearch\Model\AppUUID;
@@ -119,6 +120,7 @@ abstract class ApisearchServerBundleFunctionalTest extends BaseFunctionalTest
         $bundles = [
             BaseBundle::class,
             ApisearchServerBundle::class,
+            ApisearchBundle::class,
             ElasticaPluginBundle::class,
             ApisearchPluginsBundle::class,
         ];
@@ -162,29 +164,6 @@ abstract class ApisearchServerBundleFunctionalTest extends BaseFunctionalTest
             ],
             'apisearch' => [
                 'repositories' => [
-                    'main' => [
-                        'adapter' => 'service',
-                        'endpoint' => '~',
-                        'app_id' => self::$appId,
-                        'token' => '~',
-                        'test' => true,
-                        'search' => [
-                            'repository_service' => 'apisearch_server.items_repository',
-                        ],
-                        'app' => [
-                            'repository_service' => 'apisearch_server.app_repository',
-                        ],
-                        'user' => [
-                            'repository_service' => 'apisearch.user_repository_mock',
-                        ],
-                        'indices' => [
-                            self::$index => self::$index,
-                            self::$anotherIndex => self::$anotherIndex,
-                            $composedIndex => $composedIndex,
-                            '*' => '*',
-                            self::$yetAnotherIndex => self::$yetAnotherIndex,
-                        ],
-                    ],
                     'search_http' => [
                         'adapter' => 'http_test',
                         'endpoint' => '~',
